Function buscaminas::writer::write_data

fn write_data(data: &Vec<Vec<Cell>>, file: &mut File) -> Result<(), GameError>

Receives a Cell matrix and an opened valid File. Translates the Cell field into chars, ‘*’ for bombs and ‘.’ for empty cells and writes it in the file.

Errors

This function will return an error if it could not write the file correctly (GameError::CoulNotWriteFile).